+++ src/components/RecentProduct/RecentProduct.jsx
@@ -1,4 +1,4 @@
-import { useContext} from 'react'
+import { useContext, useMemo} from 'react'
 import { Link } from 'react-router-dom';
 import { CartContext } from '../../../Context/CartContext';
 import toast from 'react-hot-toast';
@@ -6,28 +6,36 @@ import useProduct from '../Hooks/useProduct';
 import { WishListContext } from '../../../Context/WishListContext';
 import useCategory from '../Hooks/useCategory';
 export default function RecentProduct() {
+
+  
+  
   const { addToWishList, wish ,deleteFromWish } = useContext(WishListContext);
   const { addToCart } = useContext(CartContext);
-  async function addWish(id) {
-    let { data } = await addToWishList(id);
-   
-      toast.loading('Adding Product To WishList');
-      setTimeout(()=>{
-        if (data?.status == 'success') {
+  const isProductInWishlist = useMemo(
+    () => (id) => wish?.data?.some((product) => product.id === id),
+    [wish]
+  );
+
+  async function handleWishlist(id) {
+    if (isProductInWishlist(id)) {
+      let { data } = await deleteFromWish(id);
+      toast.loading("Removing Product From WishList");
+      setTimeout(() => {
+        if (data?.status === "success") {
           toast.dismiss();
           toast.success(data?.message);
         }
-      },800)
-    }
-    async function deleteWish(id) {
-      let { data } = await deleteFromWish(id);
-      toast.loading('Removing Product From WishList');
-      setTimeout(()=>{
-        if (data?.status == 'success') {
+      }, 800);
+    } else {
+      let { data } = await addToWishList(id);
+      toast.loading("Adding Product To WishList");
+      setTimeout(() => {
+        if (data?.status === "success") {
           toast.dismiss();
-          toast.success(data?.message)
+          toast.success(data?.message);
         }
-      },800)
+      }, 800);
+    }
   }
   
   async function addCart(id) {
@@ -41,6 +49,7 @@ export default function RecentProduct() {
     },800)
   }
   let { data, isLoading } = useProduct();
+  let products =  useMemo(()=> data?.data?.data || [] , [data]);
   let x = useCategory();
   if (isLoading || x.isLoading) {
     return  <div className="flex justify-center min-h-screen fixed top-0 left-0 right-0 bottom-0 bg-white items-center">
@@ -51,19 +60,17 @@ export default function RecentProduct() {
   return (
     <>
       <div className="grid grid-cols-12 gap-5 w-[90%] m-auto">
-        {data?.data?.data?.map((product) =>
+        {products.map((product) =>
           <div key={product.id} className="relative group col-span-6 md:col-span-4 lg:col-span-3 xl:col-span-2  overflow-hidden  hover:scale-[1.059] duration-300 rounded-lg shadow-lg p-3 my-2 border-gray-400">
             <Link to={`productDetailes/${product.id}/${product.category.name}`}>
               <img loading='lazy' src={product.imageCover} className='w-full' alt={product.title} />
             </Link>
-            {(wish?.data != "")? wish?.data?.map((products) => ( products.id == product.id) ? 
-              <button  onClick={() => deleteWish(product.id)} className='z-10 absolute group-hover:right-6  transition-all duration-[.4s] top-6 -right-10 p-0 bg-transparent'><i  className="fa-solid fa-heart text-2xl  text-green-400" /> </button>
-              :
-              <button onClick={() => addWish(product.id)}  className='absolute group-hover:right-6  transition-all duration-[.4s] top-6 -right-10 p-0 bg-transparent'><i className="fa-regular fa-heart text-2xl  text-green-400" />  </button>
-            )
-            :
-            <button onClick={() => addWish(product.id)} className='absolute group-hover:right-6  transition-all duration-[.4s] top-6 -right-10 p-0 bg-transparent'><i className="fa-regular fa-heart text-2xl  text-green-400" />  </button>
-            }
+            <button
+              onClick={() => handleWishlist(product.id)}
+              className="absolute group-hover:right-6 transition-all duration-[.4s] top-6 -right-10 p-0 bg-transparent"
+            >
+              <i className={`fa-${isProductInWishlist(product.id) ? "solid" : "regular"} fa-heart text-2xl text-green-400`} />
+            </button>
             <h3 className='text-green-400 text-left px-2'  >{product.category.name}</h3>
             <h4 className='text-left px-2'>{product.title.split(' ').slice(0, 2).join(' ')}</h4>
             <div className="flex justify-between">
